Add category container to reuse in outfitForm
