+++ src/components/Closet/Closet.js
@@ -1,21 +1,11 @@
-import { v4 as uuidv4 } from 'uuid';
-import { Link } from 'react-router-dom';
+import CategoryContainer from '../CategoryContainer/CategoryContainer'
 import './Closet.css'
 
 const Closet = ({closeMenu}) => {
-  const categories = ['Pants', 'Skirts', 'Tops', 'One Pieces', 'Outer Wear', 'Bags', 'Shoes', 'Accessories', 'Miscellaneous']
-  const categoryEls = categories.map(category => {
-    const id = uuidv4();
-    const linkLocation = category.split(' ').join('').toLowerCase()
-    return <Link to={`/closet/${linkLocation}`} className='closet-link category' key={id} id={id} onClick={() => closeMenu('close')}><p className='category-text'>{category}</p></Link>
-  })
-
   return (
     <section className='closet-page'>
       <h1 className='page-title' >My Closet</h1>
-      <div className='category-container'>
-        {categoryEls}
-      </div>
+      <CategoryContainer closeMenu={closeMenu} parentRoute={'closet' }/>
     </section>
   )
 }
