+++ src/components/Outfits/Outfits.js
@@ -1,6 +1,13 @@
-const Outfits = () => {
+import add from '../../images/add.png'
+import { Link } from 'react-router-dom'
+
+const Outfits = ({closeMenu}) => {
+
   return (
-    <p>Outfits will go here</p>
+    <>
+      <Link to='/outfitform' className='menu-btn add-btn' onClick={() => closeMenu('close')}><img src={add} alt='close button' /></Link>
+      <p>Outfits will go here</p>
+    </>
   )
 }
