+++ src/index.css
@@ -49,3 +49,4 @@ code {
   align-items: center;
   justify-content: center;
 }
+
